refactor: create dukan ui states
